finished refactoring code
